Update Config module CSS and page component
This commit updates the Config.module.css file to set the width to 100% and adds missing newline at the end of the file. It also updates the page.tsx file to add select options for teams, a contest code input, and a switch for isLive. Additionally, it modifies the theme.ts file to change the list background and color.